Check jar file

+++ tools/developer_tools/python-client/generatePyClient.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-
+set -ex
 # Copyright (c) UChicago Argonne, LLC. All rights reserved.
 # See LICENSE file.
 
@@ -31,7 +31,7 @@ CDB_OPENAPI_YML_URL="$1/$CDB_OPENAPI_YML_PATH"
 
 cd $ROOT_DIR
 
-curl -O $OPEN_API_GENERATOR_JAR_URL
+[ ! -f "${OPEN_API_GENERATOR_JAR}" ] && curl -O $OPEN_API_GENERATOR_JAR_URL
 
 java -jar $OPEN_API_GENERATOR_JAR  generate -i "$CDB_OPENAPI_YML_URL" -g python -o $GEN_OUT_DIR -c $GEN_CONFIG_FILE_PATH
